feat: implement project rendering and websocket server
Adds project rendering functionality using ffmpeg and a websocket
server for communication. Also fixes timeline clip position type.

+++ Shared/src/main/java/org/turbomedia/turboedit/shared/project/TimelineClip.java
@@ -1,4 +1,4 @@
 package org.turbomedia.turboedit.shared.project;
 
-public record TimelineClip(int file, int position, long startTime, long endTime) {
+public record TimelineClip(int file, long position, long startTime, long endTime) {
 }
